+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/ModifyUserReqBody.java
@@ -1,0 +1,7 @@
+package com.ll.TeamProject.domain.user.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record ModifyUserReqBody (
+        @NotBlank String nickname
+) {}
