+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/VerificationCodeVerifyRequest.java
@@ -1,4 +1,4 @@
-package com.ll.TeamProject.domain.user.dto.admin;
+package com.ll.TeamProject.domain.user.dto;
 
 import lombok.NonNull;
