+++ backend/src/main/java/com/ll/TeamProject/domain/user/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.ll.TeamProject.domain.user.service;
 
 import com.ll.TeamProject.domain.user.dto.UserDto;
-import com.ll.TeamProject.domain.user.dto.admin.LoginDto;
+import com.ll.TeamProject.domain.user.dto.LoginDto;
 import com.ll.TeamProject.domain.user.entity.SiteUser;
 import com.ll.TeamProject.domain.user.repository.UserRepository;
 import com.ll.TeamProject.global.exceptions.ServiceException;
